refactor: begin separating load graph and geom graph functions
