+++ src/papermodels/datatypes/load_graph.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 from typing import Optional
+import math
 from copy import deepcopy
 import networkx as nx
 import hashlib
+
+from shapely import Point,
+
 from papermodels.datatypes.element import (
     Element,
-    element_to_beam_model,
-    element_to_joist_model,
     get_tag_type,
     get_normalized_coordinate,
 )
@@ -123,3 +125,60 @@ class LoadGraph(nx.DiGraph):
                 pass
 
 
+
+
+def element_to_beam_model(element: Element) -> FEModel3D:
+    """
+    Returns an FEModel3D for the data in 'element'
+    """
+    model = FEModel3D()
+    model.add_material("default", 1, 1, 1, 1)
+    element_length = element.geometry.length
+    start_point = Point(element.geometry.coords[0])
+    add_start_node = True
+    add_end_node = True
+    start_node = "N0"
+    model.add_node(start_node, 0, 0, 0)
+    for node_name, point in element.intersections:
+        x_coord = start_point.distance(point)
+        if math.isclose(x_coord, element_length, abs_tol=1e-3):
+            last_node = node_name
+            add_end_node = False
+        if math.isclose(x_coord, 0, abs_tol=1e-3):
+            model.Nodes.pop(start_node)
+            start_node = node_name
+        model.add_node(node_name, x_coord, 0, 0)
+        model.def_support(node_name, support_DY=True)
+    model.def_support(node_name, 1, 1, 1, 1, 1, 0)
+
+    if add_end_node:
+        last_node = "Nn"
+        model.add_node(last_node, element_length, 0, 0)
+    model.add_member(element.tag, "N0", last_node, "default", 1, 1, 1, 1)
+    model.add_load_combo("Pass", {"pass": 1.0})
+    return model
+
+
+def element_to_joist_model(element: Element, w: float = 0.) -> Joist:
+    """
+    Returns a Joist object based on the data in 'element'
+    """
+    try:
+        r1, r2 = element.intersections
+    except ValueError:
+        raise ValueError(f"Joists currently need to have two supports. {element.tag=} | {element.intersections=}")
+    elem_geometry = element.geometry
+    i_end = Point(elem_geometry.coords[0])
+    j_end = Point(elem_geometry.coords[1])
+    length = elem_geometry.length
+    
+    # R1 should be closest to I-end
+    r1_geom = r1[1]
+    r2_geom = r2[1]
+    if i_end.distance(r1_geom) > i_end.distance(r2_geom):
+        r1, r2 = r2, r1
+    
+    span = r1_geom.distance(r2_geom)
+    a_cantilever = round(abs(i_end.distance(r1_geom)), 6)
+    b_cantilever = round(abs(r2_geom.distance(j_end)), 6)
+    return Joist(span, a_cantilever, b_cantilever)
